+++ R/imports.R
@@ -2,3 +2,6 @@
 #' @import methods
 #' @noRd
 "_PACKAGE"
+
+methods::setOldClass("DistributionModel")
+methods::setOldClass("BiodiversityScenario")